Breakout tests of functions to individual python files under subdirectories
Removes returning user definition array in db_utils.py; opts to have the user manage username directly via parameter. If specific userid is passed, return that within the same if-branch that sets it, instead of doing a second if-branch to see if the parameter is set; reverts a misunderstanding between username and sysname, insofar that dashes are permitted in usernames, but stripped in sysnames

Merges tests from 'test_utilities.py' into 'test_password_secure.py', and deletes 'test_utilities.py'

Style fixes for added tests

Fixes to accomodate flake8 implemented throughout the added tests
